+++ src/main/java/wooteco/subway/auth/ui/AuthController.java
@@ -36,19 +36,4 @@ public class AuthController {
         return ResponseEntity.ok().build();
     }
 
-    @ExceptionHandler(InvalidMemberException.class)
-    private ResponseEntity<Void> handleInvalidMemberException(Exception e) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-    }
-
-    @ExceptionHandler(Exception.class)
-    private ResponseEntity<Void> handleException(Exception e) {
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-    }
-
-    @ExceptionHandler(InvalidTokenException.class)
-    private ResponseEntity<Void> handleInvalidTokenException(Exception e) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
-    }
-
 }
